+++ apps/scoutgame/components/[username]/PublicProfilePage.tsx
