+++ apps/scoutgame/components/[username]/components/PublicBuilderProfile/PublicBuilderProfileContainer.tsx
@@ -7,8 +7,8 @@ import { BuilderCard } from 'components/common/Card/BuilderCard/BuilderCard';
 import type { ScoutInfo } from 'components/common/Card/ScoutCard';
 import { ScoutsGallery } from 'components/common/Gallery/ScoutsGallery';
 import { UserProfile } from 'components/common/Profile/UserProfile';
-import { BuilderActivitiesList } from 'components/profile/mine/components/BuilderProfile/BuilderActivitiesList';
-import { BuilderWeeklyStats } from 'components/profile/mine/components/BuilderProfile/BuilderWeeklyStats';
+import { BuilderActivitiesList } from 'components/profile/components/BuilderProfile/BuilderActivitiesList';
+import { BuilderWeeklyStats } from 'components/profile/components/BuilderProfile/BuilderWeeklyStats';
 import { useMdScreen } from 'hooks/useMediaScreens';
 import type { BuilderActivity } from 'lib/builders/getBuilderActivities';
 import type { BasicUserInfo } from 'lib/users/interfaces';
