+++ apps/scoutgame/components/[username]/components/PublicScoutProfile/PublicScoutProfileStats.tsx
