+++ apps/scoutgame/components/profile/components/PointsClaimScreen/PointsClaimButton.tsx
