+++ apps/scoutgame/components/profile/components/PointsClaimScreen/PointsClaimScreen.tsx
