+++ apps/scoutgame/components/profile/components/PointsClaimScreen/QualifiedActionsTable.tsx
