+++ apps/scoutgame/components/profile/components/ScoutProfile/ScoutStats.tsx
